requirements.txt
beautifulsoup4
cryptography
filetype
prompt_toolkit
pypdf
python-dotenv
requests
